Moved incursion functions to incursion.js
