+++ commands/sentry/initializeembed.js
@@ -15,5 +15,6 @@ module.exports = {
       message.channel.send(newEmbed).then(message => {
         message.channel.send("Server ID: " + message.guild.id + "\nChannel ID: " + message.channel.id + "\nMessage ID: " + message.id)
       })
+      message.react("✅")
     }
 }
