+++ commands/sentry/repairscomplete.js
@@ -3,8 +3,7 @@ module.exports = {
 	description: 'Changes repair target to a standby message',
 	usage: '',
 	permlvl: 1, // 0 = Everyone, 1 = Mentor, 2 = Staff
-	execute(message, args, passArray) {
-		updateEmbedField = passArray[4]
+	execute(message, args, updateEmbedField) {
 		updateEmbedField({ name: "**Repairs:**", value: "All repairs completed. Please stand by for further targets."})
 		message.react("✅")
 	},
